Use decimal for price and double for stock amount

+++ POSeidon/Product.cs
@@ -4,8 +4,8 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public bool IsCountable { get; set; }
-        public int StockAmount { get; set; }
+        public double StockAmount { get; set; }
     }
 }
